Remove the call to entry.setTime when building the JarEntry. When running against OpenJDK 10 this causes an exception to occur. Throws java.util.zip.ZipException: invalid CEN header (bad header size)
setTime seems to be more/less the same as setLastModifiedTime so I don't know that we need both, but this worked in Oracle Java 1.8 and prior. I have submitted a bug to Oracle, and JarOutputStreamBugReport demonstrates the issue.

We may want to update this code if we hear back from Oracle on the bug report.